circular imports and typing
